requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
moviepy @ git+https://github.com/Zulko/moviepy.git@f515b100e1fd069da6f29d5b01b1d42b52d8068d
narwhals==1.40.0
numpy<2
packaging==24.2
pandas==2.2.3
pillow==11.2.1
proglog==0.1.12
protobuf==6.31.0
pyarrow==20.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
scikit-learn
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yt-dlp==2025.5.22
openai-whisper
torch==2.2.2
librosa==0.10.1

